# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Crafted By API",
        "version": "0.1"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get a list of products",
                "description": "Returns list of products",
                "operationId": "getProductList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductCard"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Products Store",
                "description": "Create Product",
                "operationId": "addProduct",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductModel"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductModel"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "405": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/products/{product_id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Find a product by Id",
                "description": "Return a single product",
                "operationId": "getProductById",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "path",
                        "description": "Id of product to return",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductModel"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        },
        "/products/{id}": {
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Update existing product",
                "description": "Returns updated product data",
                "operationId": "updateProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product id",
                        "required": true,
                        "schema": {
                            "type": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductModel"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductModel"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Delete existing product",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteProduct",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product id",
                        "required": true,
                        "schema": {
                            "type": "uuid"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "BusinessCard": {
                "title": "Business Card",
                "description": "Small Business Data",
                "properties": {
                    "Id": {
                        "title": "ID",
                        "description": "ID",
                        "format": "uuid"
                    },
                    "name": {
                        "title": "Business name",
                        "type": "string"
                    },
                    "email": {
                        "title": "Business email",
                        "type": "string"
                    },
                    "theme": {
                        "$ref": "#/components/schemas/Theme"
                    }
                },
                "type": "object"
            },
...
            "ProductCard": {
                "title": "ProductCard",
                "description": "Product Model for cards",
                "properties": {
                    "Id": {
                        "title": "ID",
                        "description": "ID",
                        "format": "uuid"
                    },
                    "name": {
                        "title": "Name",
                        "description": "Product name",
                        "type": "string",
                        "example": "Chaise en bois rare"
                    },
                    "image": {
                        "title": "Image",
                        "description": "Image URL",
                        "type": "string"
                    },
                    "price": {
                        "title": "Price",
                        "description": "Product price",
                        "type": "number",
                        "format": "float",
                        "example": "123.45"
                    },
                    "active": {
                        "title": "Active",
                        "description": "Is product available to sell",
                        "type": "boolean"
                    },
                    "business": {
                        "$ref": "#/components/schemas/BusinessCard"
                    }
                },
                "type": "object"
            },
            "ProductModel": {
                "title": "Product Model",
                "properties": {
                    "Id": {
                        "title": "ID",
                        "description": "ID",
                        "format": "uuid"
                    },
                    "name": {
                        "title": "Product name",
                        "description": "Product name",
                        "type": "string",
                        "example": "Chaise en bois rare"
                    },
                    "image": {
                        "title": "Product image",
                        "description": "Image URL",
                        "type": "string"
                    },
                    "price": {
                        "title": "Product price",
                        "description": "Product price",
                        "type": "number",
                        "format": "float",
                        "example": "123.45"
                    },
                    "active": {
                        "title": "Is active",
                        "description": "Is product available to sell",
                        "type": "boolean"
                    },
                    "business": {
                        "$ref": "#/components/schemas/BusinessCard"
                    },
                    "description": {
                        "title": "Product description",
                        "description": "Product description",
                        "type": "string",
                        "example": "Chaise en bois rare d'Indonésie taillé à la main."
                    },
                    "stock_quantity": {
                        "title": "Product stock quantity",
                        "description": "Quantity of product available",
                        "type": "integer",
                        "example": "7"
                    },
                    "height": {
                        "title": "Product height",
                        "type": "string"
                    },
                    "width": {
                        "title": "Product width",
                        "type": "string"
                    },
                    "depth": {
                        "title": "Product depth",
                        "type": "string"
                    },
                    "capacity": {
                        "title": "product capacity",
                        "type": "string"
                    },
                    "category": {
                        "title": "Product category",
                        "type": "string"
                    },
                    "material": {
                        "title": "Product material",
                        "type": "string"
                    },
                    "style": {
                        "title": "Product style",
                        "type": "string"
                    },
                    "color": {
                        "title": "Product color",
                        "type": "string"
                    },
                    "created_at": {
                        "title": "Product registery date",
                        "type": "string",
                        "format": "date"
                    }
                },
                "type": "object"
            },
            "Theme": {
                "title": "Business Theme",
                "properties": {
                    "layer": {
                        "title": "Theme Layer",
                        "type": "string"
                    },
                    "color1": {
                        "title": "Theme color n°1",
                        "type": "string"
                    },
                    "color2": {
                        "title": "Theme color n°2",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            },
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}